refactor: refactoring code and fixing regex
