+++ packages/jsActions/native-mobile-resources/src/deeplink/ParseUrlToObject.ts
@@ -5,7 +5,7 @@
 // - the code between BEGIN USER CODE and END USER CODE
 // - the code between BEGIN EXTRA CODE and END EXTRA CODE
 // Other code you write will be lost the next time you deploy the project.
-import * as UrlParse from "url-parse";
+import UrlParse from "url-parse";
 
 type Records = Record<string, string | undefined>;
